requirements.txt
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.57.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
lxml==5.4.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
onnxruntime==1.21.1
opencv-python-headless==4.11.0.86
packaging==25.0
pdf2docx==0.5.8
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==6.30.2
PyMatting==1.1.13
PyMuPDF==1.25.5
pyreadline3==3.5.4
python-docx==1.1.2
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
sympy==1.13.3
termcolor==3.0.1
tifffile==2025.3.30
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
